+++ gensim/models/experimental/__init__.py
@@ -1,3 +1,6 @@
 from .dssm import DSSM  # noqa:F401
-from .drmm_tks import DRMM_TKS  # noqa:F401
 from .drmm import DRMM  # noqa:F401
+from .custom_losses import rank_hinge_loss  # noqa:F401
+from .custom_callbacks import ValidationCallback  # noqa:F401
+from .evaluation_metrics import mapk, mean_ndcg  # noqa:F401
+from .drmm_tks import DRMM_TKS_Model  # noqa:F401
