+++ gensim/models/experimental/custom_callbacks.py
