--- gensim/similarity_learning/HowToRunModels.md
@@ -1,23 +1,0 @@
-# Running Similarity Learning Models
-
-## DRMM_TKS (Deep Relevance Matching Model Top k Solutions)
-Example usage:
-
-For evaluating drmm_tks on the WikiQA corpus
-$ python evaluate_models.py --model drmm_tks --datapath ../data/WikiQACorpus/ --word_embedding_path ../evaluation_scripts/glove.6B.50d.txt --result_save_path results_drmm_tks
-
-
-**Link to Paper:** [A Deep Relevance Matching Model for Ad-hoc Retrieval](http://www.bigdatalab.ac.cn/~gjf/papers/2016/CIKM2016a_guo.pdf)
-
-Refer to the image below to understand the network topology. <br/>Note: The image boxes are labelled so:
-`a = Dense(10)(b) -> a_Dense_10_b`
-
-![DRMM_TKS_Topology](model.png)
-
-## DSSM (Deep Structured Semantic Model)
-Example Usage:
-`$ python dssm_example.py --wikiqa_folder_path ./data/WikiQACorpus/`
-
-Note: There's a bug in the WikiQAExtractor for DSSM. Might lead to minor problems. Will be fixed/reimplemented when we get to that model.
-
-**Link to Paper:** [Learning Deep Structured Semantic Models for Web Search using Clickthrough Data](https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/cikm2013_DSSM_fullversion.pdf)
