--- gensim/similarity_learning/__init__.py
@@ -1,4 +1,0 @@
-from .custom_losses import rank_hinge_loss  # noqa:F401
-from .custom_callbacks import ValidationCallback  # noqa:F401
-from .evaluation_metrics import mapk, mean_ndcg  # noqa:F401
-from .DRMM_TKS_Model import DRMM_TKS_Model  # noqa:F401
