--- gensim/similarity_learning/preprocessing/__init__.py
@@ -1,2 +1,0 @@
-from .sl_vocab import WikiQAExtractor, WikiQA_DRMM_TKS_Extractor  # noqa:F401
-from .list_generator import ListGenerator  # noqa:F401
